+++ src/blob/BlobServerFactory.ts
@@ -4,6 +4,7 @@ import { DEFAULT_SQL_OPTIONS } from "../common/utils/constants";
 import BlobConfiguration from "./BlobConfiguration";
 import BlobEnvironment from "./BlobEnvironment";
 import BlobServer from "./BlobServer";
+import IBlobEnvironment from "./IBlobEnvironment";
 import SqlBlobConfiguration from "./SqlBlobConfiguration";
 import SqlBlobServer from "./SqlBlobServer";
 import { DEFAULT_BLOB_PERSISTENCE_PATH } from "./utils/constants";
@@ -14,14 +15,22 @@ import {
 } from "./utils/constants";
 
 export class BlobServerFactory {
-  public async createServer(): Promise<BlobServer | SqlBlobServer> {
+  public async createServer(
+    blobEnvironment?: IBlobEnvironment
+  ): Promise<BlobServer | SqlBlobServer> {
     // TODO: Check it's in Visual Studio Code environment or not
     const isVSC = false;
 
     if (!isVSC) {
-      const env = new BlobEnvironment();
+      const env = blobEnvironment ? blobEnvironment : new BlobEnvironment();
       const location = await env.location();
       const debugFilePath = await env.debug();
+
+      if (typeof debugFilePath === "boolean") {
+        throw RangeError(
+          `Must provide a debug log file path for parameter -d or --debug`
+        );
+      }
 
       DEFAULT_BLOB_PERSISTENCE_ARRAY[0].locationPath = join(
         location,
@@ -42,7 +51,8 @@ export class BlobServerFactory {
           !env.silent(),
           undefined,
           debugFilePath !== undefined,
-          debugFilePath
+          debugFilePath,
+          env.loose()
         );
 
         return new SqlBlobServer(config);
@@ -56,7 +66,8 @@ export class BlobServerFactory {
           !env.silent(),
           undefined,
           debugFilePath !== undefined,
-          debugFilePath
+          debugFilePath,
+          env.loose()
         );
         return new BlobServer(config);
       }